+++ api/https
@@ -22,7 +22,7 @@
 
 #include "net/http/response.hpp"
 #include "net/http/request.hpp"
-#include "net/botan/http_server.hpp"
-#include "net/openssl/server.hpp"
+#include "net/https/botan_server.hpp"
+#include "net/https/openssl_server.hpp"
 
 #endif
